Update F&S modifications to work with u-boot-2024.04
Some changes are necessary to get F&S boards running again:

- Link Time Optimization (CONFIG_LTO) was automatically selected when
  CONFIG_IMX8M was active. However LTO is not possible for F&S DRAM timing
  files for NBoot. Disable LTO on F&S boards for now.
- Function is_imx8m_running_secondary_boot_image() is now officially available
  as boot_mode_getprosec(), however the prototype was missing.
- Fix print_hex_dump() function when using DUMP_PREFIX_OFFSET.
- In include/firmware/imx/sci/sci.h, add missing prototype for function
  sc_misc_board_ioctl().
- Fix spinand_manufacturers[] entry for ESMT chips, the name has changed.
- Minor modifications in F&S board files and common files, e.g. include some
  more header files or change function signatures that have changed in the
  meantime.
- Remove setup_fec() and setup_eqos() from fsimx8mp.c, this is handled in
  clock_imx8mm.c now.
- Update all F&S defconfigs, e.g. change CONFIG_SYS_SPL_MALLOC_* to
  CONFIG_SPL_SYS_MALLOC_*, add CMD_NFS which is now off by default, and keep
  CONFIG_TOOLS_LIBCRYPTO off.
- Only build splash.c if CONFIG_SPLASH_SCREEN is set.

+++ cmd/update.c
@@ -58,7 +58,7 @@ static int update_ram(const char *action, const char **check,
 #ifdef CONFIG_CMD_UBIFS
 extern int cmd_ubifs_mount(const char *vol_name);
 extern int set_ubi_part(const char *part_name, const char *vid_header_offset);
-extern int ubifs_load(const char *filename, u32 addr, u32 size);
+extern int ubifs_load(const char *filename, unsigned long addr, u32 size);
 
 static int update_ubi(const char *action, const char *part_name,
 		      const char *vol_name, const char *fname,
